可以从OLED添加指纹

+++ src/fingermain.h
@@ -5,7 +5,7 @@
 void handle_interrupt1();
 void handle_interrupt2();
 void handle_interrupt3();
-uint8_t fingermain_init();
+int fingermain_init();
 uint8_t fingermain_run();
 void fingerTask(void *parameter);
 #endif
