+++ src/fingerprint.h
@@ -2,7 +2,7 @@
 #define FINGERPRINT_HH
 #include <Arduino.h>
 
-uint8_t Fingerprint_inti();
+int Fingerprint_inti();
 uint8_t getFingerprintID();
 int getFingerprintIDez();
 #endif
